update chat route & small fixes
