+++ app/api/chat/[id]/route.ts
@@ -1,9 +1,17 @@
 import { NextResponse } from "next/server";
-import { getChatById, getMessagesByChatId } from "@/lib/db/queries";
-import { Message } from "ai";
 import { getIronSession } from "iron-session";
 import { SessionData, sessionOptions } from "@/lib/session";
 import { cookies } from "next/headers";
+import {
+  getChatById,
+  getMessagesByChatId,
+  saveMessage,
+} from "@/lib/db/queries";
+import { createDataStreamResponse, Message, streamText } from "ai";
+import { Chat } from "@/lib/db/schema";
+import { generateSystemPrompt } from "../../prompt";
+import { isAddress } from "viem";
+import { groqModel } from "@/lib/ai/groq";
 
 type MessageContent = Message & { content: { response: string; news: [] } };
 
@@ -46,5 +54,81 @@ export async function GET(
       content: JSON.stringify(content.response),
       annotations: content.news ? [{ news: content.news }] : undefined,
     })),
+  });
+}
+
+export async function PUT(
+  request: Request,
+  { params }: { params: Promise<{ id: string }> }
+) {
+  const { id } = await params;
+  const req = await request.json();
+  const { messages }: { messages: Message[]; marketId: string } = req;
+  const userMessages = messages.filter((message) => message.role === "user");
+  const lastUserMessage = userMessages.at(-1);
+  const session = await getIronSession<SessionData>(
+    await cookies(),
+    sessionOptions
+  );
+
+  if (!session || !session.userId) {
+    return new Response("Unauthorized", { status: 401 });
+  }
+
+  const chat: Chat = await getChatById({ id });
+
+  if (!chat) {
+    return new Response("Chat not found", { status: 404 });
+  }
+
+  if (chat.userId !== session.userId) {
+    return new Response("Unauthorized", { status: 401 });
+  }
+
+  if (!lastUserMessage) {
+    return new Response("User message not found", { status: 400 });
+  }
+
+  const address =
+    chat?.marketAddress && isAddress(chat.marketAddress)
+      ? chat.marketAddress.toLowerCase()
+      : null;
+
+  const systemPrompt = await generateSystemPrompt({
+    marketAddress: address,
+  });
+
+  if (userMessages.length > 1) {
+    await saveMessage({
+      chatId: chat.id,
+      message: { response: lastUserMessage.content as string },
+      role: "user",
+    });
+  }
+
+  return createDataStreamResponse({
+    execute: (dataStream) => {
+      const result = streamText({
+        model: groqModel,
+        messages,
+        system: systemPrompt,
+        onFinish: async ({ response }) => {
+          try {
+            await saveMessage({
+              chatId: chat.id,
+              message: {
+                response: (response.messages[0].content[0] as { text: string })
+                  .text,
+              },
+              role: "assistant",
+            });
+          } catch (error) {
+            console.error("Failed to save chat:", error);
+          }
+        },
+      });
+
+      result.mergeIntoDataStream(dataStream);
+    },
   });
 }
